Update and fix various unit tests that used glTF 1.0 manifest fragments as input
